// app/api/types.ts
export interface ApiResponse<T = void> {
  success: boolean;
  data?: T;
  error?: string;
}
// ...
export type MedicineLogResponse = Omit<
  MedicineLog,
  "time" | "createdAt" | "updatedAt" | "deletedAt"
> & {
  time: string;
  createdAt: string;
  updatedAt: string;
  deletedAt: string | null;
};

export interface MedicineLogCreate {
  babyId: string;
  medicineId: string;
  time: string;
  doseAmount: number;
  unitAbbr?: string | null;
  notes?: string;
}